requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.5
uvicorn==0.31.0
paho-mqtt==1.6.1
python-multipart==0.0.9
jinja2==3.1.4
pillow==10.4.0
numpy==2.1.3
transformers==4.44.2

torch==2.4.1+cpu
scikit-learn==1.5.2
pandas==2.2.3
# Optional for best accuracy (enable if you choose --model xgb)
xgboost==2.1.2
